// src/app/modules/SSLCommerz/sslCommerz.interface.ts
export interface ISSLCommerz {
  amount: number;
  transactionId: string;
  name: string;
  email: string;
  phoneNumber: string;
  address: string;
  //   SSL_SUCCESS_BACKEND_URL: string;
  //   SSL_FAIL_BACKEND_URL: string;
  //   SSL_CANCEL_BACKEND_URL: string;
  //   SSL_SUCCESS_FRONTEND_URL: string;
  //   SSL_FAIL_FRONTEND_URL: string;
  //   SSL_CANCEL_FRONTEND_URL: string;
}